Made changes to REST UnJoin
Added new functionality to Join/Unjoin.  This new functionality will
allow the user to specify more details on how to route the two devices

XMS REST interface was changed to require the ID of the endpoint in the
UnJoin.  Thus added the ability to pass in either the ID or Call Object
into unjoin to allow for this.  Additionally, added in saving of the
last joined componant to allow for backwards compatibility for unjoin
with no id.   However please use with case as multiple unjoins/joins or
adding/removing from conference may prevent this from working as
intended.   Feel free to report any state issue found with this new
feature.